GraphMultiEPG: set default change bouquet true
If we don't offer this action by default, people don't know about it, if we press in GraphMultiEPG or GraphMultiEPGList the keys CH+- we will change the bouquet in the graph.
If it is disabled, the logical thing to do and the guideline is to zap is to use UP DOWN to move through the events until you change the next page, but you will not use these buttons for any other purpose.
Now we are going to indicate it on this screen so that people can use it by default and know what its purpose is.
Add to Skin https://github.com/OpenVisionE2/OctEtFHD-skin/commit/cf3d7cdaa0728f70c010e409e7f9a26da9ddd736