Timeshift: New option for available space on device
if there is space equal to or less than 200 MB on the device, it will not perform timeshift.
We have two ways to notify it, in the Timeshift screen a footnote is added and we have a message notification if this condition is true where it will warn us that there is little disk space for timeshift.
MovieSelection -Timeshift is included to create timeshift directories (user settings on restore backups), if these directories are different than default they will be created, from here the paths are already set if the devices are mounted, otherwise Otherwise, the pathStatus handler will tell us what type of error exists.