Timeshift: Create directories for USB and HDD
We reduce code, to do the same function.
this code executes the creation of a user-allocated timeshift directory ( backup restore ) if it does not exist.

With this, a directory is always assigned for timeshift, either the default directory or a directory previously created by the user in copy restoration, the system remembers it and creates it again if it does not exist.

If the timeshift directory is now different from the restore backup directory, pathStatus assigns it as a flash directory -> Timeshift cannot be executed.

The default folder for timeshift will now be "recordings" it is the final path where timeshifts are stored since timeshift makes recordings.
This code makes sense:
There are many manufacturers that manufacture boxes without internal hdd, small boxes is the increasingly common pattern, boxes with USB ports to assign a storage device... these boxes by default if we add a storage HW it will detect it as USB and not as HDD.

So with this code we evaluate two path possibilities for timeshift, the old HDD that is kept and we give the possibility if there is only one default path as a mount point ( USB ) it will be able to assign default paths to both directories by default, before only to to HDD.

If we have a default storage device as mount point /media/usb with the old code, the user has to bother to assign the timeshift path even creating it, with this code we cover two common mount points USB and HDD and it would not have to create absolutely nothing to do timeshift.