FallbackTimer.py: show error with NotificationWithID
RuntimeError: [StartEnigma] Error: Modal starts are only allowed from a screen that is modal!

This problem is generated in Python 3, there are errors in the TimerList XML open and read functions, when we import timers from the server, it is still to be solved, there are no execution problems in Python 2.

.- Indentation is corrected.