[HarddiskSetup] Change Initialize to Format
A storage device is initialised automatically when you record, go to flash online etc without any user interevention.
The context of the menu/ question is that it is asking if we want to format the device, not if we want to initialise it. (initail setup for use).
Most E2 users are from a Windows background and Format in this context reflects the function in a more user friendly term.

Co-authored-by: AbuBaniaz <AbuBaniaz@users.noreply.github.com>
Co-authored-by: norhap <elfogondesegovia@gmail.com>